requirements.txt
aiogram>=3.7,<4.0
aiohttp>=3.9
python-dotenv>=1.0.0
flyerapi
